+++ android/nativebrik/src/main/java/com/nativebrik/sdk/sdk.kt
@@ -34,7 +34,7 @@ import kotlin.time.Duration
 import kotlin.time.DurationUnit
 import kotlin.time.toDuration
 
-const val VERSION = "0.4.14"
+const val VERSION = "0.5.0"
 
 data class Endpoint(
     val cdn: String = "https://cdn.nativebrik.com",